📍 Feat(#66): connect socket and linking roomId
